smalloc
Implement simple versions of malloc and free called smalloc and sfree.
These two functions will operate nearly the same way as malloc and
free: smalloc makes memory available and sfree frees that memory. One
difference between free and your sfree is that sfree will return a -1
on error and a 0 on success.